Change myf to back to double to fix compiler error.
git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@9912 0381ead4-6506-0410-b988-94b70fbc4730

+++ littrack/base/CbmLitFloat.h
@@ -11,6 +11,6 @@
 #ifndef CBMLITFLOAT_H_
 #define CBMLITFLOAT_H_
 
-typedef float myf;
+typedef double myf;
 
 #endif /*CBMLITFLOAT_H_*/
